The changes in this commit are as listed below.     - Added a HTTP Client Wrapper Class initialization method.     - Added missing PHP doc-blocks.
This commit just contains the above mentioned minor changes. The first change is adding
the class "\AdityaZanjad\Http\Http". This class is a like an entry-point for the
package. It will allow its users to pick the "HTTP Client Wrapper Class",
initialize it & then, send a HTTP request. The second commit just
adds some missing doc-blocks to the PHP code.